requirements.txt
asgiref==3.5.2
Django==4.1.1
django-environ==0.9.0
django-phonenumber-field==7.0.0
djangorestframework==3.13.1
phonenumbers==8.12.56
psycopg2==2.9.3
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
